Implement ArrayAccess on all data collectors for further back-compat with third party plugins.
