+++ data/block_editor/post_block.php
@@ -10,7 +10,12 @@
  * @package query-monitor
  */
 
-class QM_Data_Post_Block {
+/**
+ * @implements \ArrayAccess<string,mixed>
+ */
+class QM_Data_Post_Block implements \ArrayAccess {
+	use QM_ArrayAccess;
+
 	/**
 	 * @var string|null
 	 */
